requirements.txt
numpy==2.1.2
pillow==11.0.0